requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[argon2]
python-jose
python-dotenv
pdfplumber
pytesseract
Pillow
openai
requests
google-auth
httpx
celery
redis
jobspy
xhtml2pdf
jinja2
anthropic
